+++ assignment_250618_BookInfo/Resources/Assets.xcassets/AppIcon.appiconset/Contents.json
