+++ assignment_250618_BookInfo/Resources/Assets.xcassets/Contents.json
