+++ assignment_250618_BookInfo/Resources/Base.lproj/LaunchScreen.storyboard
