+++ assignment_250618_BookInfo/Resources/Images.xcassets/Contents.json
