+++ assignment_250618_BookInfo/Resources/Images.xcassets/harrypotter1.imageset/Contents.json
