+++ assignment_250618_BookInfo/Resources/Images.xcassets/harrypotter2.imageset/Contents.json
