+++ assignment_250618_BookInfo/Resources/Images.xcassets/harrypotter5.imageset/Contents.json
