+++ assignment_250618_BookInfo/Resources/Images.xcassets/harrypotter7.imageset/Contents.json
