+++ assignment_250618_BookInfo/Resources/data.json
