+++ assignment_250618_BookInfo/Services/DataService.swift
@@ -7,29 +7,6 @@
 
 import Foundation
 
-struct BookResponse: Codable {
-    let data: [BookData]
-}
-
-struct BookData: Codable {
-    let attributes: Book
-}
-
-struct Book: Codable {
-    var title: String
-    var author: String
-    var pages: Int
-    var release_date: String
-    var dedication: String
-    var summary: String
-    var wiki: String
-    var chapters: [Chapter]
-}
-
-struct Chapter: Codable {
-    var title: String
-}
-
 class DataService {
     
     enum DataError: Error {
